Splits index.js into different controllers. Splits config into a standalone file.
