+++ controllers/logoutController.js
@@ -1,0 +1,3 @@
+module.exports.get = function(req, res) {
+  res.clearCookie('token').redirect('/login');
+}
